requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.63.2
pinecone-client==3.0.0
twilio==8.12.0
requests==2.31.0
geopy==2.4.1
firebase-admin==6.4.0
elevenlabs==0.2.26
rich==13.7.0 